+++ examples/todo/src/1_schema.ts
@@ -58,7 +58,7 @@ export const TodoAccount = co
     /** The account migration is run on account creation and on every log-in.
      *  You can use it to set up the account root and any other initial CoValues you need.
      */
-    if (!account.root) {
+    if (account.root === undefined) {
       account.root = TodoAccountRoot.create({
         projects: co.list(TodoProject).create([], { owner: account }),
       });