+++ packages/jazz-tools/src/react-native-core/index.ts
@@ -1,7 +1,7 @@
-export * from "./provider.js";
 export * from "./auth/auth.js";
 export * from "./hooks.js";
 export * from "./media.js";
+export * from "./provider.js";
 export * from "./storage/kv-store-context.js";
 
 export { SQLiteDatabaseDriverAsync } from "cojson-storage";