+++ packages/jazz-tools/src/tools/coValues/group.ts
@@ -182,7 +182,9 @@ export class Group extends CoValueBase implements CoValue {
     }
   }
 
-  get members(): Array<{
+  private getMembersFromKeys(
+    accountIDs: Iterable<RawAccountID | AgentID>,
+  ): Array<{
     id: string;
     role: AccountRole;
     ref: Ref<Account>;
@@ -195,7 +197,7 @@ export class Group extends CoValueBase implements CoValue {
       optional: false,
     } satisfies RefEncoded<Account>;
 
-    for (const accountID of this._raw.getAllMemberKeysSet()) {
+    for (const accountID of accountIDs) {
       if (!isAccountID(accountID)) continue;
 
       const role = this._raw.roleOf(accountID);
@@ -228,6 +230,30 @@ export class Group extends CoValueBase implements CoValue {
     }
 
     return members;
+  }
+
+  /**
+   * Returns all members of the group, including inherited members from parent
+   * groups.
+   *
+   * If you need only the direct members of the group, use
+   * {@link getDirectMembers} instead.
+   *
+   * @returns The members of the group.
+   */
+  get members() {
+    return this.getMembersFromKeys(this._raw.getAllMemberKeysSet());
+  }
+
+  /**
+   * Returns the direct members of the group.
+   *
+   * If you need all members of the group, including inherited members from
+   * parent groups, use {@link Group.members|members} instead.
+   * @returns The direct members of the group.
+   */
+  getDirectMembers() {
+    return this.getMembersFromKeys(this._raw.getMemberKeys());
   }
 
   getRoleOf(member: Everyone | ID<Account> | "me") {